% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_cerebro_v1.4.R
\name{Cerebro_v1.4}
\alias{Cerebro_v1.4}
\title{R6 class in which data sets will be stored for visualization in Cerebro.}
\value{
A new \code{Cerebro_v1.4} object.
}
\description{
A \code{Cerebro_v1.4} object is an R6 class that contains several types of
data that can be visualized in Cerebro.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{version}}{cerebroApp version that was used to create the object.}

\item{\code{experiment}}{\code{list} that contains meta data about the data set,
including experiment name, species, date of export.}

\item{\code{technical_info}}{\code{list} that contains technical information
about the analysis, including the R session info.}

\item{\code{parameters}}{\code{list} that contains important parameters that
were used during the analysis, e.g. cut-off values for cell filtering.}

\item{\code{groups}}{\code{list} that contains specified grouping variables and
and the group levels (subgroups) that belong to each of them. For each
grouping variable, a corresponding column with the same name must exist
in the meta data.}

\item{\code{cell_cycle}}{\code{vector} that contains the name of columns in the
meta data that contain cell cycle assignments.}

\item{\code{gene_lists}}{\code{list} that contains gene lists, e.g.
mitochondrial and/or ribosomal genes.}

\item{\code{expression}}{\code{matrix}-like object that holds transcript counts.}

\item{\code{meta_data}}{\code{data.frame} that contains cell meta data.}

\item{\code{projections}}{\code{list} that contains projections/dimensional
reductions.}

\item{\code{most_expressed_genes}}{\code{list} that contains a \code{data.frame}
holding the most expressed genes for each grouping variable that was
specified during the call to \code{\link{getMostExpressedGenes}}.}

\item{\code{marker_genes}}{\code{list} that contains a \code{list} for every
method that was used to calculate marker genes, and a \code{data.frame}
for each grouping variable, e.g. those that were specified during the
call to \code{\link{getMarkerGenes}}.}

\item{\code{enriched_pathways}}{\code{list} that contains a \code{list} for
every method that was used to calculate marker genes, and a
\code{data.frame} for each grouping variable, e.g. those that were
specified during the call to \code{\link{getEnrichedPathways}} or
\code{\link{performGeneSetEnrichmentAnalysis}}.}

\item{\code{trees}}{\code{list} that contains a phylogenetic tree (class
\code{phylo}) for grouping variables.}

\item{\code{trajectories}}{\code{list} that contains a \code{list} for every
method that was used to calculate trajectories, and, depending on the
method, a \code{data.frame} or \code{list} for each specific trajectory,
e.g. those extracted with \code{\link{extractMonocleTrajectory}}.}

\item{\code{extra_material}}{\code{list} that can contain additional material
related to the data set; tables should be stored in \code{data.frame}
format in a named \code{list} called `tables`}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Cerebro_v1.4-new}{\code{Cerebro_v1.4$new()}}
\item \href{#method-Cerebro_v1.4-setVersion}{\code{Cerebro_v1.4$setVersion()}}
\item \href{#method-Cerebro_v1.4-getVersion}{\code{Cerebro_v1.4$getVersion()}}
\item \href{#method-Cerebro_v1.4-checkIfGroupExists}{\code{Cerebro_v1.4$checkIfGroupExists()}}
\item \href{#method-Cerebro_v1.4-checkIfColumnExistsInMetadata}{\code{Cerebro_v1.4$checkIfColumnExistsInMetadata()}}
\item \href{#method-Cerebro_v1.4-addExperiment}{\code{Cerebro_v1.4$addExperiment()}}
\item \href{#method-Cerebro_v1.4-getExperiment}{\code{Cerebro_v1.4$getExperiment()}}
\item \href{#method-Cerebro_v1.4-addParameters}{\code{Cerebro_v1.4$addParameters()}}
\item \href{#method-Cerebro_v1.4-getParameters}{\code{Cerebro_v1.4$getParameters()}}
\item \href{#method-Cerebro_v1.4-addTechnicalInfo}{\code{Cerebro_v1.4$addTechnicalInfo()}}
\item \href{#method-Cerebro_v1.4-getTechnicalInfo}{\code{Cerebro_v1.4$getTechnicalInfo()}}
\item \href{#method-Cerebro_v1.4-addGroup}{\code{Cerebro_v1.4$addGroup()}}
\item \href{#method-Cerebro_v1.4-getGroups}{\code{Cerebro_v1.4$getGroups()}}
\item \href{#method-Cerebro_v1.4-getGroupLevels}{\code{Cerebro_v1.4$getGroupLevels()}}
\item \href{#method-Cerebro_v1.4-setMetaData}{\code{Cerebro_v1.4$setMetaData()}}
\item \href{#method-Cerebro_v1.4-getMetaData}{\code{Cerebro_v1.4$getMetaData()}}
\item \href{#method-Cerebro_v1.4-addGeneList}{\code{Cerebro_v1.4$addGeneList()}}
\item \href{#method-Cerebro_v1.4-getGeneLists}{\code{Cerebro_v1.4$getGeneLists()}}
\item \href{#method-Cerebro_v1.4-setExpression}{\code{Cerebro_v1.4$setExpression()}}
\item \href{#method-Cerebro_v1.4-getCellNames}{\code{Cerebro_v1.4$getCellNames()}}
\item \href{#method-Cerebro_v1.4-getGeneNames}{\code{Cerebro_v1.4$getGeneNames()}}
\item \href{#method-Cerebro_v1.4-getMeanExpressionForGenes}{\code{Cerebro_v1.4$getMeanExpressionForGenes()}}
\item \href{#method-Cerebro_v1.4-getMeanExpressionForCells}{\code{Cerebro_v1.4$getMeanExpressionForCells()}}
\item \href{#method-Cerebro_v1.4-getExpressionMatrix}{\code{Cerebro_v1.4$getExpressionMatrix()}}
\item \href{#method-Cerebro_v1.4-setCellCycle}{\code{Cerebro_v1.4$setCellCycle()}}
\item \href{#method-Cerebro_v1.4-getCellCycle}{\code{Cerebro_v1.4$getCellCycle()}}
\item \href{#method-Cerebro_v1.4-addProjection}{\code{Cerebro_v1.4$addProjection()}}
\item \href{#method-Cerebro_v1.4-availableProjections}{\code{Cerebro_v1.4$availableProjections()}}
\item \href{#method-Cerebro_v1.4-getProjection}{\code{Cerebro_v1.4$getProjection()}}
\item \href{#method-Cerebro_v1.4-addTree}{\code{Cerebro_v1.4$addTree()}}
\item \href{#method-Cerebro_v1.4-getTree}{\code{Cerebro_v1.4$getTree()}}
\item \href{#method-Cerebro_v1.4-addMostExpressedGenes}{\code{Cerebro_v1.4$addMostExpressedGenes()}}
\item \href{#method-Cerebro_v1.4-getGroupsWithMostExpressedGenes}{\code{Cerebro_v1.4$getGroupsWithMostExpressedGenes()}}
\item \href{#method-Cerebro_v1.4-getMostExpressedGenes}{\code{Cerebro_v1.4$getMostExpressedGenes()}}
\item \href{#method-Cerebro_v1.4-addMarkerGenes}{\code{Cerebro_v1.4$addMarkerGenes()}}
\item \href{#method-Cerebro_v1.4-getMethodsForMarkerGenes}{\code{Cerebro_v1.4$getMethodsForMarkerGenes()}}
\item \href{#method-Cerebro_v1.4-getGroupsWithMarkerGenes}{\code{Cerebro_v1.4$getGroupsWithMarkerGenes()}}
\item \href{#method-Cerebro_v1.4-getMarkerGenes}{\code{Cerebro_v1.4$getMarkerGenes()}}
\item \href{#method-Cerebro_v1.4-addEnrichedPathways}{\code{Cerebro_v1.4$addEnrichedPathways()}}
\item \href{#method-Cerebro_v1.4-getMethodsForEnrichedPathways}{\code{Cerebro_v1.4$getMethodsForEnrichedPathways()}}
\item \href{#method-Cerebro_v1.4-getGroupsWithEnrichedPathways}{\code{Cerebro_v1.4$getGroupsWithEnrichedPathways()}}
\item \href{#method-Cerebro_v1.4-getEnrichedPathways}{\code{Cerebro_v1.4$getEnrichedPathways()}}
\item \href{#method-Cerebro_v1.4-addTrajectory}{\code{Cerebro_v1.4$addTrajectory()}}
\item \href{#method-Cerebro_v1.4-getMethodsForTrajectories}{\code{Cerebro_v1.4$getMethodsForTrajectories()}}
\item \href{#method-Cerebro_v1.4-getNamesOfTrajectories}{\code{Cerebro_v1.4$getNamesOfTrajectories()}}
\item \href{#method-Cerebro_v1.4-getTrajectory}{\code{Cerebro_v1.4$getTrajectory()}}
\item \href{#method-Cerebro_v1.4-addExtraMaterial}{\code{Cerebro_v1.4$addExtraMaterial()}}
\item \href{#method-Cerebro_v1.4-addExtraTable}{\code{Cerebro_v1.4$addExtraTable()}}
\item \href{#method-Cerebro_v1.4-addExtraPlot}{\code{Cerebro_v1.4$addExtraPlot()}}
\item \href{#method-Cerebro_v1.4-getExtraMaterialCategories}{\code{Cerebro_v1.4$getExtraMaterialCategories()}}
\item \href{#method-Cerebro_v1.4-checkForExtraTables}{\code{Cerebro_v1.4$checkForExtraTables()}}
\item \href{#method-Cerebro_v1.4-getNamesOfExtraTables}{\code{Cerebro_v1.4$getNamesOfExtraTables()}}
\item \href{#method-Cerebro_v1.4-getExtraTable}{\code{Cerebro_v1.4$getExtraTable()}}
\item \href{#method-Cerebro_v1.4-checkForExtraPlots}{\code{Cerebro_v1.4$checkForExtraPlots()}}
\item \href{#method-Cerebro_v1.4-getNamesOfExtraPlots}{\code{Cerebro_v1.4$getNamesOfExtraPlots()}}
\item \href{#method-Cerebro_v1.4-getExtraPlot}{\code{Cerebro_v1.4$getExtraPlot()}}
\item \href{#method-Cerebro_v1.4-print}{\code{Cerebro_v1.4$print()}}
\item \href{#method-Cerebro_v1.4-clone}{\code{Cerebro_v1.4$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-new"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Cerebro_v1.4} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Cerebro_v1.4} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-setVersion"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-setVersion}{}}}
\subsection{Method \code{setVersion()}}{
Set the version of \code{cerebroApp} that was used to generate this
object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$setVersion(version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{Version to set.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getVersion"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getVersion}{}}}
\subsection{Method \code{getVersion()}}{
Get the version of \code{cerebroApp} that was used to generate this
object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getVersion()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Version as \code{package_version} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-checkIfGroupExists"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-checkIfGroupExists}{}}}
\subsection{Method \code{checkIfGroupExists()}}{
Safety function that will check if a provided group name is present in
the \code{groups} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$checkIfGroupExists(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name to be tested}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-checkIfColumnExistsInMetadata"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-checkIfColumnExistsInMetadata}{}}}
\subsection{Method \code{checkIfColumnExistsInMetadata()}}{
Safety function that will check if a provided group name is present in
the meta data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$checkIfColumnExistsInMetadata(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name to be tested.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-addExperiment"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-addExperiment}{}}}
\subsection{Method \code{addExperiment()}}{
Add information to \code{experiment} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$addExperiment(field, content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{Name of the information, e.g. \code{organism}.}

\item{\code{content}}{Actual information, e.g. \code{hg}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getExperiment"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getExperiment}{}}}
\subsection{Method \code{getExperiment()}}{
Retrieve information from \code{experiment} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getExperiment()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of all entries in the \code{experiment} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-addParameters"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-addParameters}{}}}
\subsection{Method \code{addParameters()}}{
Add information to \code{parameters} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$addParameters(field, content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{Name of the information, e.g. \code{number_of_PCs}.}

\item{\code{content}}{Actual information, e.g. \code{30}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getParameters"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getParameters}{}}}
\subsection{Method \code{getParameters()}}{
Retrieve information from \code{parameters} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getParameters()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of all entries in the \code{parameters} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-addTechnicalInfo"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-addTechnicalInfo}{}}}
\subsection{Method \code{addTechnicalInfo()}}{
Add information to \code{technical_info} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$addTechnicalInfo(field, content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{Name of the information, e.g. \code{R}.}

\item{\code{content}}{Actual information, e.g. \code{4.0.2}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getTechnicalInfo"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getTechnicalInfo}{}}}
\subsection{Method \code{getTechnicalInfo()}}{
Retrieve information from \code{technical_info} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getTechnicalInfo()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of all entries in the \code{technical_info} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-addGroup"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-addGroup}{}}}
\subsection{Method \code{addGroup()}}{
Add group to the groups registered in the \code{groups} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$addGroup(group_name, levels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name.}

\item{\code{levels}}{\code{vector} of group levels (subgroups).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getGroups"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getGroups}{}}}
\subsection{Method \code{getGroups()}}{
Retrieve all names in the \code{groups} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getGroups()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of registered groups.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getGroupLevels"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getGroupLevels}{}}}
\subsection{Method \code{getGroupLevels()}}{
Retrieve group levels for a group registered in the \code{groups} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getGroupLevels(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name for which to retrieve group levels.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{vector} of group levels.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-setMetaData"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-setMetaData}{}}}
\subsection{Method \code{setMetaData()}}{
Set meta data for cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$setMetaData(table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table}}{\code{data.frame} that contains meta data for cells. The
number of rows must be equal to the number of rows of projections and
the number of columns in the transcript count matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getMetaData"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getMetaData}{}}}
\subsection{Method \code{getMetaData()}}{
Retrieve meta data for cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getMetaData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{data.frame} containing meta data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-addGeneList"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-addGeneList}{}}}
\subsection{Method \code{addGeneList()}}{
Add a gene list to the \code{gene_lists}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$addGeneList(name, genes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the gene list.}

\item{\code{genes}}{\code{vector} of genes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getGeneLists"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getGeneLists}{}}}
\subsection{Method \code{getGeneLists()}}{
Retrieve gene lists from the \code{gene_lists}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getGeneLists()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of all entries in the \code{gene_lists} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-setExpression"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-setExpression}{}}}
\subsection{Method \code{setExpression()}}{
Set transcript count matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$setExpression(counts)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{counts}}{\code{matrix}-like object that contains transcript counts
for cells in the data set. Number of columns must be equal to the number
of rows in the \code{meta_data} field.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getCellNames"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getCellNames}{}}}
\subsection{Method \code{getCellNames()}}{
Get names of all cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getCellNames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} containing all cell names/barcodes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getGeneNames"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getGeneNames}{}}}
\subsection{Method \code{getGeneNames()}}{
Get names of all genes in transcript count matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getGeneNames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} containing all gene names in transcript count matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getMeanExpressionForGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getMeanExpressionForGenes}{}}}
\subsection{Method \code{getMeanExpressionForGenes()}}{
Retrieve mean expression across all cells in the data set for a set of
genes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getMeanExpressionForGenes(genes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Names of genes to extract; no default.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} containing specified gene names and their respective
mean expression across all cells in the data set.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getMeanExpressionForCells"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getMeanExpressionForCells}{}}}
\subsection{Method \code{getMeanExpressionForCells()}}{
Retrieve (mean) expression for a single gene or a set of genes for a
given set of cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getMeanExpressionForCells(cells = NULL, genes = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cells}}{Names/barcodes of cells to extract; defaults to \code{NULL},
which will return all cells.}

\item{\code{genes}}{Names of genes to extract; defaults to \code{NULL}, which
will return all genes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{vector} containing (mean) expression across all specified genes in
each specified cell.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getExpressionMatrix"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getExpressionMatrix}{}}}
\subsection{Method \code{getExpressionMatrix()}}{
Retrieve transcript count matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getExpressionMatrix(cells = NULL, genes = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cells}}{Names/barcodes of cells to extract; defaults to \code{NULL},
which will return all cells.}

\item{\code{genes}}{Names of genes to extract; defaults to \code{NULL}, which
will return all genes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Dense transcript count matrix for specified cells and genes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-setCellCycle"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-setCellCycle}{}}}
\subsection{Method \code{setCellCycle()}}{
Add columns containing cell cycle assignments to the \code{cell_cycle}
field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$setCellCycle(cols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cols}}{\code{vector} of columns names containing cell cycle
assignments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getCellCycle"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getCellCycle}{}}}
\subsection{Method \code{getCellCycle()}}{
Retrieve column names containing cell cycle assignments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getCellCycle()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of column names in meta data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-addProjection"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-addProjection}{}}}
\subsection{Method \code{addProjection()}}{
Add projections (dimensional reductions).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$addProjection(name, projection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the projection.}

\item{\code{projection}}{\code{data.frame} containing positions of cells in
projection.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-availableProjections"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-availableProjections}{}}}
\subsection{Method \code{availableProjections()}}{
Get list of available projections (dimensional reductions).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$availableProjections()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of projections / dimensional reductions that are available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getProjection"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getProjection}{}}}
\subsection{Method \code{getProjection()}}{
Retrieve data for a specific projection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getProjection(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of projection.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} containing the positions of cells in the projection.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-addTree"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-addTree}{}}}
\subsection{Method \code{addTree()}}{
Add phylogenetic tree to \code{trees} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$addTree(group_name, tree)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name that this tree belongs to.}

\item{\code{tree}}{Phylogenetic tree as \code{phylo} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getTree"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getTree}{}}}
\subsection{Method \code{getTree()}}{
Retrieve phylogenetic tree for a specific group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getTree(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name for which to retrieve phylogenetic tree.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Phylogenetic tree as \code{phylo} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-addMostExpressedGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-addMostExpressedGenes}{}}}
\subsection{Method \code{addMostExpressedGenes()}}{
Add table of most expressed genes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$addMostExpressedGenes(group_name, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Name of grouping variable that the most expressed genes
belong to. Must be registered in the \code{groups} field.}

\item{\code{table}}{\code{data.frame} that contains the most expressed genes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getGroupsWithMostExpressedGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getGroupsWithMostExpressedGenes}{}}}
\subsection{Method \code{getGroupsWithMostExpressedGenes()}}{
Retrieve names of grouping variables for which most expressed genes are
available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getGroupsWithMostExpressedGenes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of grouping variables for which most expressed genes are
available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getMostExpressedGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getMostExpressedGenes}{}}}
\subsection{Method \code{getMostExpressedGenes()}}{
Retrieve table of most expressed genes for a grouping variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getMostExpressedGenes(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Grouping variable for which most expressed genes should
be retrieved.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} that contains most expressed genes for group levels of
the specified grouping variable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-addMarkerGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-addMarkerGenes}{}}}
\subsection{Method \code{addMarkerGenes()}}{
Add table of marker genes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$addMarkerGenes(method, name, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method that was used to generate the marker genes.}

\item{\code{name}}{Name of table. This name will be used to select the table in
Cerebro. It is recommended to use the grouping variable, e.g.
\code{sample}.}

\item{\code{table}}{\code{data.frame} that contains the marker genes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getMethodsForMarkerGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getMethodsForMarkerGenes}{}}}
\subsection{Method \code{getMethodsForMarkerGenes()}}{
Retrieve names of methods that were used to generate marker genes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getMethodsForMarkerGenes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of names of methods that were used to generate marker
genes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getGroupsWithMarkerGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getGroupsWithMarkerGenes}{}}}
\subsection{Method \code{getGroupsWithMarkerGenes()}}{
Retrieve grouping variables for which marker genes were generated using
a specified method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getGroupsWithMarkerGenes(method)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{vector} of grouping variables for which marker genes were
calculated using the specified method.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getMarkerGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getMarkerGenes}{}}}
\subsection{Method \code{getMarkerGenes()}}{
Retrieve table of marker genes for specific method and grouping variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getMarkerGenes(method, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method.}

\item{\code{name}}{Name of table.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} that contains marker genes for the specified
combination of method and grouping variable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-addEnrichedPathways"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-addEnrichedPathways}{}}}
\subsection{Method \code{addEnrichedPathways()}}{
Add table of enriched pathways.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$addEnrichedPathways(method, name, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method that was used to generate the enriched
pathways.}

\item{\code{name}}{Name of table. This name will be used to select the table in
Cerebro. It is recommended to use the grouping variable, e.g.
\code{sample}.}

\item{\code{table}}{\code{data.frame} that contains the enriched pathways.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getMethodsForEnrichedPathways"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getMethodsForEnrichedPathways}{}}}
\subsection{Method \code{getMethodsForEnrichedPathways()}}{
Retrieve names of methods that were used to generate enriched pathways.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getMethodsForEnrichedPathways()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of names of methods that were used to generate enriched
pathways.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getGroupsWithEnrichedPathways"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getGroupsWithEnrichedPathways}{}}}
\subsection{Method \code{getGroupsWithEnrichedPathways()}}{
Retrieve grouping variables for which enriched pathways were generated
using a specified method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getGroupsWithEnrichedPathways(method)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{vector} of grouping variables for which enriched pathways were
calculated using the specified method.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getEnrichedPathways"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getEnrichedPathways}{}}}
\subsection{Method \code{getEnrichedPathways()}}{
Retrieve table of enriched pathways for specific method and grouping
variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getEnrichedPathways(method, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method.}

\item{\code{name}}{Grouping variable.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} that contains enriched pathways for the specified
combination of method and grouping variable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-addTrajectory"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-addTrajectory}{}}}
\subsection{Method \code{addTrajectory()}}{
Add trajectory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$addTrajectory(method, name, content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method that was used to generate the trajectory.}

\item{\code{name}}{Name of the trajectory. This name will be used later in
Cerebro to select the trajectory.}

\item{\code{content}}{Relevant data for the trajectory, depending on the method
this could be a \code{list} holding edges, cell positions, pseudotime,
etc.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getMethodsForTrajectories"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getMethodsForTrajectories}{}}}
\subsection{Method \code{getMethodsForTrajectories()}}{
Retrieve names of methods that were used to generate trajectories.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getMethodsForTrajectories()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of names of methods that were used to generate
trajectories.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getNamesOfTrajectories"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getNamesOfTrajectories}{}}}
\subsection{Method \code{getNamesOfTrajectories()}}{
Retrieve names of available trajectories for a specified method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getNamesOfTrajectories(method)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{vector} of available trajectory for the specified method.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getTrajectory"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getTrajectory}{}}}
\subsection{Method \code{getTrajectory()}}{
Retrieve data for a specific trajectory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getTrajectory(method, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method.}

\item{\code{name}}{Name of trajectory.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The type of data depends on the method that was used to generate the
trajectory.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-addExtraMaterial"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-addExtraMaterial}{}}}
\subsection{Method \code{addExtraMaterial()}}{
Add content to extra material field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$addExtraMaterial(category, name, content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{category}}{Name of category. At the moment, only \code{tables} and
\code{plots} are valid categories. Tables must be in \code{data.frame}
format and plots must be created with \code{ggplot2}.}

\item{\code{name}}{Name of material, will be used to select it in Cerebro.}

\item{\code{content}}{Data that should be added.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-addExtraTable"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-addExtraTable}{}}}
\subsection{Method \code{addExtraTable()}}{
Add table to `extra_material` slot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$addExtraTable(name, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of material, will be used to select it in Cerebro.}

\item{\code{table}}{Table that should be added, must be \code{data.frame}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-addExtraPlot"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-addExtraPlot}{}}}
\subsection{Method \code{addExtraPlot()}}{
Add plot to `extra_material` slot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$addExtraPlot(name, plot)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of material, will be used to select it in Cerebro.}

\item{\code{plot}}{Plot that should be added, must be created with
\code{ggplot2} (class: \code{ggplot}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getExtraMaterialCategories"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getExtraMaterialCategories}{}}}
\subsection{Method \code{getExtraMaterialCategories()}}{
Get names of categories for which extra material is available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getExtraMaterialCategories()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} with names of available categories.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-checkForExtraTables"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-checkForExtraTables}{}}}
\subsection{Method \code{checkForExtraTables()}}{
Check whether there are tables in the extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$checkForExtraTables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} indicating whether there are tables in the extra
materials.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getNamesOfExtraTables"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getNamesOfExtraTables}{}}}
\subsection{Method \code{getNamesOfExtraTables()}}{
Get names of tables in extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getNamesOfExtraTables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} containing names of tables in extra materials.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getExtraTable"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getExtraTable}{}}}
\subsection{Method \code{getExtraTable()}}{
Get table from extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getExtraTable(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of table.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Requested table in \code{data.frame} format.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-checkForExtraPlots"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-checkForExtraPlots}{}}}
\subsection{Method \code{checkForExtraPlots()}}{
Check whether there are plots in the extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$checkForExtraPlots()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} indicating whether there are plots in the extra
materials.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getNamesOfExtraPlots"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getNamesOfExtraPlots}{}}}
\subsection{Method \code{getNamesOfExtraPlots()}}{
Get names of plots in extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getNamesOfExtraPlots()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} containing names of plots in extra materials.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-getExtraPlot"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-getExtraPlot}{}}}
\subsection{Method \code{getExtraPlot()}}{
Get plot from extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$getExtraPlot(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of plot.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Requested plot made with \code{ggplot2}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-print"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-print}{}}}
\subsection{Method \code{print()}}{
Show overview of object and the data it contains.
Print overview of available marker gene results for \code{self$print()}
function.
Print overview of available enriched pathway results for
\code{self$print()} function.
Print overview of available trajectories for \code{self$print()} function.
Print overview of extra material for \code{self$print()} function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.4-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.4-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.4$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
